refactor: combine FeedTransformation and TypedFeedTransformation
